SpotifyController/requirements.txt
python-dotenv==1.0.1

spotipy==2.24.0